Generate Doctrine entities for comments and tags
Steps:
 - run 'app/console doctrine:generate:entity'
   - Entity shortcut name: TngWorkshopBoardBundle:BoardMessage
   - Fields: userName/string/255, date/datetime, messageText/text
   - generate repository class: yes
 - run again
   - Entity shortcut name: TngWorkshopBoardBundle:BoardTag
   - Fields: tag/string/255
   - generate repository class: yes

Note:
Alternatively, you can run
  app/console doctrine:mapping:import "TngWorkshopBoardBundle" annotation
to read the schema from the existing database, and then refactor the
generated entities to fit your taste. Then run
  doctrine:generate:entities
to generate accessors.
Note that we don't need a mapping for the tags_comments table; this will
be taken care of in the next step.